BLD: move exhaustive benchmarks to separate module, add option to include that module
